neuroslice-platform/api-dashboard-tier/dashboard-backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
PyJWT==2.10.1
pydantic==2.11.4
